requirements.txt
# Core FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Browser automation
browser-use>=0.1.0
playwright>=1.40.0
selenium>=4.15.0

# AI/LLM providers
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
mistralai>=0.0.12
ollama>=0.1.7
boto3>=1.34.0  # For Amazon Bedrock

# Azure OpenAI (use openai package with azure endpoint)
# azure-openai>=1.0.0  # Not needed - use openai package instead

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Storage and persistence
aiosqlite>=0.19.0
sqlalchemy>=2.0.23
alembic>=1.13.0

# Media and file handling
Pillow>=10.1.0
aiofiles>=23.2.1

# HTTP client for external requests
httpx>=0.25.2
aiohttp>=3.9.1

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-json-logger>=2.0.7
structlog>=23.2.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Production deployment
gunicorn>=21.2.0
redis>=5.0.1
celery>=5.3.4